feat: separate TasksList and TasksListItem
